% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_logic.R
\name{derive_features_from_counts}
\alias{derive_features_from_counts}
\title{Derive Features from Normalized Counts}
\usage{
derive_features_from_counts(
  norm_counts,
  coldata,
  contrast_group,
  control_level
)
}
\description{
Derive Features from Normalized Counts
}
