% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_logic.R
\name{run_prioritization}
\alias{run_prioritization}
\title{Run AutoRel Prioritization}
\usage{
run_prioritization(
  norm_counts,
  res,
  coldata,
  contrast_group,
  control_level,
  model_path,
  output_path
)
}
\description{
Run AutoRel Prioritization
}
