% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate and Clean AutoRel Inputs}
\usage{
validate_inputs(norm_counts, coldata, contrast_group, control_level)
}
\arguments{
\item{norm_counts}{Dataframe of normalized counts}

\item{coldata}{Dataframe of metadata}

\item{contrast_group}{Name of the contrast column}

\item{control_level}{Name of the control group}
}
\description{
Validate and Clean AutoRel Inputs
}
